Graciously handle backends with no need for resizing, instead of throwing unimplemented!()
